adds performance test-case and removes dead code
